requirements.txt
fastapi==0.110.0
python-dotenv==1.0.1
uvicorn==0.28.0
